# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Medynium API",
    "description": "Governed Patient 360 and clinical agent on Snowflake. Decision support on synthetic data only. Errors use {error, message}. Sessions use HttpOnly cookies; state-changing requests need the X-Medynium-Client: web header.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/patients/{patient_id}/pins": {
      "get": {
        "tags": [
          "pins"
        ],
        "summary": "List Pins",
        "description": "Evidence the caller pinned to this patient's workspace.",
        "operationId": "list_pins_patients__patient_id__pins_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinList"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "pins"
        ],
        "summary": "Create Pin",
        "description": "Pin one evidence item from one of the caller's own answers. Repeats with the same key return the same pin.",
        "operationId": "create_pin_patients__patient_id__pins_post",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pin"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}/pins/{pin_id}": {
      "delete": {
        "tags": [
          "pins"
        ],
        "summary": "Delete Pin",
        "operationId": "delete_pin_patients__patient_id__pins__pin_id__delete",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "pin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pin Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "not_found",
          "unauthorized",
          "forbidden",
          "action_not_allowed",
          "agent_unavailable",
          "timeout",
          "conflict",
          "rate_limited",
          "invalid_request",
          "not_implemented",
          "internal_error"
        ],
        "title": "ErrorCode"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Pin": {
        "properties": {
          "pin_id": {
            "type": "string",
            "title": "Pin Id"
          },
          "answer_id": {
            "type": "string",
            "title": "Answer Id"
          },
          "evidence_id": {
            "type": "string",
            "title": "Evidence Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "pin_id",
          "answer_id",
          "evidence_id",
          "label",
          "note",
          "created_at"
        ],
        "title": "Pin"
      },
      "PinCreate": {
        "properties": {
          "answer_id": {
            "type": "string",
            "maxLength": 40,
            "minLength": 3,
            "title": "Answer Id"
          },
          "evidence_id": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Evidence Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "answer_id",
          "evidence_id"
        ],
        "title": "PinCreate"
      },
      "PinList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Pin"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PinList"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}